requirements.txt
streamlit==1.28.0
numpy==1.26.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.1.0
deep-translator==1.11.1
langdetect==1.0.9
huggingface-hub==0.14.1
